Set disposition table collection flag to on
Turn the collect_disposition_table flag to true when entering dispo
table and off when exiting. Turn the collect_disposition_table flag to
false when at the next table instead of at the table end tag, because
the sentencing row's on the disposition table have sub-tables in the
fee's column.