Collect date for each event
Collects all dates for each row and adds the date to its own list and
appends that to the disposition_table_data list. The
disposition_table_data is a two-dimensional array with each sub-list
corresponding to a row in the events table.

This is being done because the length of each row varies in the events
table. Each element in the list will be data from one row.