Implement section attribute
A statute's section is its combined chapter and section numbers. Some
chapters are four digits long, and others are two digits long. This is
the first implementation of these tests. The chapters that are four
digits long have a letter in the fourth position. These tests for the
creation of sections with chapters that are three and four characters
long. This may need to evolve to include chapters that are two digits
long once there is enough information to make that change.

This is being done to dry up the code as the section number is used
repeatedly through out the charge class.