Replace 'None' with null where applicable in search JSON response
This is a follow up from https://github.com/codeforpdx/recordexpungPDX/pull/566. Otherwise we would receive a "Nothing was returned from render. This usually means a return statement is missing. Or, to render nothing, return null" error. Eventually we should validate against a JSON schema so that we don't have these kinds of errors.