Add pytest-lazy-fixture library
This library will be used to parameterize fixtures in our round trip serialization tests. This is a highly wanted feature in pytest itself and there is a proposal to get it into the main library: https://docs.pytest.org/en/latest/proposals/parametrize_with_fixtures.html. Thus eventually I expect this library to be removed.