Refactor CaseParser.feed to return CaseParserData
This makes it clearer that we are invoking CaseParser.feed to obtain the 4 fields in CaseParserData, and nothing else. `self.parser` in `test_case_parser.py` should probably be renamed to `self.parser_data` or even `self.data` as it is the returned data that we are testing not the internal attributes of the parser itself; however, that would touch a fair bit of code (even if it is just trivial renaming), so I decided to leave it as is.