new Results Summary panel in frontend. (#756)
* new Results Summary panel in frontend.

 * The file structure here puts all the new required react components in a shared folder. That's because all of these additional components are used only in the RecordSummary panel. I'd like to reorg several of the other component files into a similar structure later.

 * This has a minor backend refactor to the record and summary data format: 'summary' is being attached to the record object which isn't awesome, but it reduces required changes to the frontend.
Basically, the Record type in the frontend gets passed to the SearchResults component which contains the Summary panel. To get the Summary data object into the Summary component without it being a subfield of record would require a bit more fiddling with redux , and/or passing adjacent props fields into the component hierarchy that I don't want to do right now. That might be a better solution but really only to make the backend have this 'record' object not get a new field added to it after it gets constructed by the expunger. 

 * renamed 'records' field to 'record' to match backend & spec.

* record.summary has a default empty RecordSummary object

* Various fixes from comments and minor improvements

* Include total balance in summary panel

* Add CaseNumber component IDs and link them in the summary panel