Add form validation to EditUser component. (#826)
* Add form validation to EditUser component.

Validation checks that name, email, and password fields exist, checks
the email is valid, checks that password is valid, and checks that
the passwords match. Password and Confirm Password fields were added
so the admin can change any or all of the user's info.

* Refactor, conditionally render role selection, hide password reset email button.

Refactor input aria-invalid attributes. Add isAdmin condition to
only allow role selection to be changed if the user is an admin.
Hide password reset email link.