Excludes a defined set of charge types (specifically, parking tickets and traffic violations) from the Summary display because they're always ineligible and they clutter the summary list. (#841)
This uses a new method on Charge type hidden_from_record_summary(), which defaults to False for most charge types, so the main work is done in the backend summarizer code.

Updates the subheader on the list of ineligible cases to clarify the new behavior. Parking violations are a subset of traffic violations, so the subheader names only the latter type.