Convert Expunger.run into a static method
This completes the refactor to make Record a frozen dataclass: since Expunger.run is now completely side-effect free, we can apply a cache over it.

Additionally adjust _without_skippable_charges so that it returns a new Record: this ensures that the removed charges are also removed from their corresponding cases. This caused a rare bug in the friendly rule implementation for juvenile charges.