Ask felony class level for a Felony Unclassified charge
This disambiguating question removes a few known instances of an Unclassified Charge Type.

It also lets us clean up Manu/Del logic a bit. It includes a small change to how manu/del charges behaved: we used to assume that, if a manu/del charge was schedule 2, or heroin, meth, or cocaine, the charge could only be Felony Class A or B. Schedule 3 or 4 could be Class A, B, or C.
Now, all of these charges with "non-marijuana identifiers" can be answered (if Felony Unclassified) as felony class A, B, or C.
This actually matches the way we handle Person Felonies: namely, it's possible that a charge that's statutorily one Felony class can be prosecuted as a different class, so we take the actual charge class that was prosecuted.