Add support for unknown/missing disposition
OECI charges will no longer have an optional Disposition. The previous value of None will become a DispositionStatus.UNKNOWN with a date of today as the disposition is unknown/missing/deferred as of today. This interpretation fits in nicely with how some existing charges have dispositions of "conditional discharge" or "deferred", which both mean that the disposition is missing as of today and will be input in the future.