Update custom PDF forms for Clackamas county
We add a new field called "case_number_with_comments". Since the Clackamas county forms do not require you to write down the charges that are to be expunged, we need to specify which charges are eligible in the case number field as a comment. For example, "Case Number 0123456 (in part, counts 1,2,4)".