Expunger checks if all other convictions are traffic related
Previously, the `_are_other_convictions_all_traffic` function was passed in
`convictions` that included **only convictions from blocking charges**. This
could lead to improper handling of records with juvenile cases.

The desired behaviour was that we want to capture **all** charges that aren't
traffic, not all **blocking** charges that aren't traffic.