Update uwsgi config to reload worker on http timeout
Setting `--http-timeout` doesn't automatically reload the workers that timed out in the background: we need to set `--harakiri` to the same length for that to happen.

Also adjust the process count to 2. I have discovered that uwsgi doesn't automatically route the request for a particular session (or even ip address) to the same worker even if that worker is available. In other words, with multiple processors there is a chance that the request will be routed to the instance of the Flask application that doesn't have the previous in-memory cache. If this becomes an issue, we should switch to using something like redis to guarantee cache integrity across instances.